requirements.txt
streamlit
opencv-python-headless
mediapipe
matplotlib
numpy
google-generativeai
Pillow